requirements.txt
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=10.0.0
psutil>=5.9.0
GPUtil>=1.4.0
requests>=2.31.0
pywin32>=306
pyautogui>=0.9.54
keyboard>=0.13.5
python-dotenv>=1.0.0
